[trail-and-admin-pages] admin-page optimize tab components
